requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
api-display-purposes==0.0.3
APScheduler==3.6.3
arrow==1.0.3
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
beautifulsoup4==4.9.3
blis==0.7.4
cachetools==4.2.2
catalogue==2.0.3
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
clarifai==2.6.2
click==7.1.2
colorama==0.4.4
configparser==3.8.1
cryptography==3.4.7
cymem==2.0.5
decorator==4.4.2
discord==1.0.1
discord.py==1.6.0
EasyProcess==0.3
emoji==1.2.0
filelock==3.0.12
fire==0.4.0
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-api-core==1.26.1
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.34.1
h11==0.12.0
h5py==3.1.0
headliner==1.0.2
httpcore==0.12.3
httplib2==0.19.0
httpx==0.17.1
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
import-ipynb==0.1.3
iniconfig==1.1.1
instabot==0.117.0
instapy==0.6.14
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lorem==0.1.1
loremipsum==1.0.5
lxml==4.6.2
Markdown==3.3.4
MarkupSafe==1.1.1
MeaningCloud-python==2.0.0
mock==4.0.3
moviepy==1.0.3
multidict==5.1.0
murmurhash==1.0.5
natsort==7.1.1
nltk==3.6.2
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.3
pathy==0.5.2
Pillow==8.1.2
pluggy==0.13.1
plyer==2.0.0
preshed==3.0.5
proglog==0.1.9
protobuf==3.15.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.4
python-dateutil==2.8.1
python-http-client==3.3.2
python-telegram-bot==13.6
pytrends==4.7.3
pytube @ git+https://github.com/nficano/pytube@8d0aaa43701d10a1235c3c7239ff88d28176e901
pytz==2021.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.13.3
rfc3986==1.4.0
rsa==4.7.2
sacremoses==0.0.45
schedule==1.0.0
scikit-learn==0.24.2
scipy==1.6.2
selenium==3.141.0
semantic-version==2.8.5
setuptools-rust==0.12.1
six==1.15.0
smart-open==3.0.0
sniffio==1.2.0
soupsieve==2.2.1
spacy==3.0.6
spacy-legacy==3.0.5
srsly==2.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc2
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
thinc==8.0.3
threadpoolctl==2.1.0
tokenizers==0.10.2
toml==0.10.2
tornado==6.1
tqdm==4.57.0
transformers==4.5.1
typer==0.3.2
typing-extensions==3.7.4.3
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.3
wasabi==0.8.2
webdriverdownloader==1.1.0.3
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
youtube-search-python==1.4.2
